move Copy to its own lang item file
