+++ crates/formality-rust/src/check/core_crate.rs
@@ -2,6 +2,7 @@ use formality_core::term;
 
 use crate::grammar::{Crate, CrateId, CrateItem};
 
+mod copy;
 mod field_projections;
 
 #[term($*items)]
@@ -12,9 +13,12 @@ struct LangItems {
 pub fn krate() -> Crate {
     Crate {
         id: CrateId::new("core"),
-        items: [field_projections::lang_items()]
-            .into_iter()
-            .flat_map(|items| items.items)
-            .collect(),
+        items: [
+            copy::lang_items(),
+            field_projections::lang_items(), //
+        ]
+        .into_iter()
+        .flat_map(|items| items.items)
+        .collect(),
     }
 }
